code_analyzer/requirements.txt
litellm
jinja2
pygments
tqdm
python-dotenv
openai
markdown
